One study per request: revamp Request Study flow
Still need to update View Request accordingly, but this is a good start.
