+++ lib/i18n/Strings.js
@@ -77,9 +77,9 @@ const REQUEST_STUDY_REQUIRES_LOCATION = {
  * @memberof Strings
  * @type {StringMessage}
  */
-const REQUEST_STUDY_REQUIRES_STUDIES = {
+const REQUEST_STUDY_REQUIRES_STUDY_TYPE = {
   variant: 'warning',
-  text: 'Please request one or more studies.',
+  text: 'Please select a study type.',
 };
 
 /**
@@ -122,7 +122,7 @@ const REQUEST_STUDY_UPDATED = {
  * @memberof Strings
  * @type {StringMessage}
  */
-const STUDY_OTHER_HOURS_REQUIRES_NOTES = {
+const REQUEST_STUDY_OTHER_HOURS_REQUIRES_NOTES = {
   variant: 'error',
   text: 'Please provide additional notes to explain requirements for Other Hours.',
 };
@@ -131,7 +131,7 @@ const STUDY_OTHER_HOURS_REQUIRES_NOTES = {
  * @memberof Strings
  * @type {StringMessage}
  */
-const STUDY_REQUIRES_DAYS_OF_WEEK = {
+const REQUEST_STUDY_REQUIRES_DAYS_OF_WEEK = {
   variant: 'error',
   text: 'Please select one or more days of the week.',
 };
@@ -144,16 +144,16 @@ const Strings = {
   OPTIONAL,
   REQUEST_STUDY_FORBIDDEN,
   REQUEST_STUDY_NOT_FOUND,
+  REQUEST_STUDY_OTHER_HOURS_REQUIRES_NOTES,
   REQUEST_STUDY_PROVIDE_URGENT_DUE_DATE,
   REQUEST_STUDY_PROVIDE_URGENT_REASON,
+  REQUEST_STUDY_REQUIRES_DAYS_OF_WEEK,
   REQUEST_STUDY_REQUIRES_LOCATION,
   REQUEST_STUDY_REQUIRES_REASONS,
-  REQUEST_STUDY_REQUIRES_STUDIES,
+  REQUEST_STUDY_REQUIRES_STUDY_TYPE,
   REQUEST_STUDY_SUBMITTED,
   REQUEST_STUDY_TIME_TO_FULFILL,
   REQUEST_STUDY_UPDATED,
-  STUDY_OTHER_HOURS_REQUIRES_NOTES,
-  STUDY_REQUIRES_DAYS_OF_WEEK,
 };
 
 export {
@@ -161,14 +161,14 @@ export {
   OPTIONAL,
   REQUEST_STUDY_FORBIDDEN,
   REQUEST_STUDY_NOT_FOUND,
+  REQUEST_STUDY_OTHER_HOURS_REQUIRES_NOTES,
   REQUEST_STUDY_PROVIDE_URGENT_DUE_DATE,
   REQUEST_STUDY_PROVIDE_URGENT_REASON,
+  REQUEST_STUDY_REQUIRES_DAYS_OF_WEEK,
   REQUEST_STUDY_REQUIRES_LOCATION,
   REQUEST_STUDY_REQUIRES_REASONS,
-  REQUEST_STUDY_REQUIRES_STUDIES,
+  REQUEST_STUDY_REQUIRES_STUDY_TYPE,
   REQUEST_STUDY_SUBMITTED,
   REQUEST_STUDY_TIME_TO_FULFILL,
   REQUEST_STUDY_UPDATED,
-  STUDY_OTHER_HOURS_REQUIRES_NOTES,
-  STUDY_REQUIRES_DAYS_OF_WEEK,
 };
