+++ lib/validation/ValidationsStudyRequest.js
@@ -51,22 +51,20 @@ const ValidationsStudyRequest = {
     reasons: {
       required,
     },
-    studies: {
+    studyType: {
       required,
-      $each: {
-        daysOfWeek: {
-          required,
-        },
-        duration: {
-          needsValidDaysOfWeek(duration, { daysOfWeek }) {
-            const k = numConsecutiveDaysOfWeek(daysOfWeek);
-            return k * 24 >= duration;
-          },
-        },
-        notes: {
-          requiredIfOtherHours: requiredIf(({ hours }) => hours === StudyHours.OTHER),
-        },
+    },
+    daysOfWeek: {
+      required,
+    },
+    duration: {
+      needsValidDaysOfWeek(duration, { daysOfWeek }) {
+        const k = numConsecutiveDaysOfWeek(daysOfWeek);
+        return k * 24 >= duration;
       },
+    },
+    notes: {
+      requiredIfOtherHours: requiredIf(({ hours }) => hours === StudyHours.OTHER),
     },
     urgentReason: {
       requiredIfUrgent(urgentReason, { urgent }) {
